requirements.txt
# Core Dependencies for AI Agents
pyyaml>=6.0
requests>=2.31.0

# LLM Providers (optional - install as needed)
openai>=1.0.0
anthropic>=0.3.0

# Additional utilities
python-dateutil>=2.8.2
pathlib2>=2.3.7; python_version < '3.4'

# Development and testing (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0